Separated the HeartsUpdateFunctions from the MainActivity class
